spline #813 Liveness and readiness probes
