+++ commons/pom.xml
@@ -96,6 +96,10 @@
             <groupId>org.scala-graph</groupId>
             <artifactId>graph-core_${scala.compat.version}</artifactId>
         </dependency>
+        <dependency>
+            <groupId>com.typesafe</groupId>
+            <artifactId>config</artifactId>
+        </dependency>
 
         <!-- test scope -->
