+++ kafka-gateway/src/main/webapp/META-INF/context.xml
@@ -1,0 +1,39 @@
+<!--
+  ~ Copyright 2019 ABSA Group Limited
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<Context>
+
+    <!--
+        Required config
+    -->
+
+    <!--
+        ArangoDB Connection URL.
+        Format: arangodb://user:password@host:port/database_name
+    -->
+    <Environment name="spline/database/connectionUrl" type="java.lang.String" override="false"/>
+
+    <Environment name="spline/kafka/topic" type="java.lang.String" override="false"/>
+    <Environment name="spline/kafka/consumer/group/id" type="java.lang.String" override="false"/>
+    <Environment name="spline/kafka/consumer/bootstrap/servers" type="java.lang.String" override="false"/>
+
+    <!--
+        Optional config
+    -->
+    <Environment name="spline/kafka/insertPlanTimeout" type="java.lang.Long" override="false"/>
+    <Environment name="spline/kafka/insertEventTimeout" type="java.lang.Long" override="false"/>
+
+</Context>
